+++ tests/testthat/test-rule-fit-binomial.R
@@ -164,3 +164,40 @@ test_that('non-formula method', {
   }
 
 })
+
+
+
+# ------------------------------------------------------------------------------
+
+test_that('tidy method - two classes', {
+  skip_on_cran()
+
+  library(xrf)
+
+  xrf_cls_mod <-
+    rule_fit(trees = 3, penalty = .001) %>%
+    set_engine("xrf") %>%
+    set_mode("classification")
+
+  set.seed(1)
+  xrf_cls_fit <-
+    xrf_cls_mod %>%
+    fit(Class ~ ., data = ad_mod)
+  xrf_rule_res <- tidy(xrf_cls_fit)
+  raw_coef <- coef(xrf_cls_fit$fit, lambda = 0.001)
+  raw_coef <- raw_coef[raw_coef[,1] != 0, ]
+  expect_true(nrow(raw_coef) == nrow(xrf_rule_res))
+  expect_true(all(raw_coef$term %in% xrf_rule_res$rule_id))
+
+  xrf_col_res <- tidy(xrf_cls_fit, unit = "column")
+  expect_equal(
+    sort(unique(xrf_col_res$term)),
+    c("(Intercept)", "Genotype", "MMP10", "p_tau")
+  )
+  expect_equal(
+    sort(unique(raw_coef$term)),
+    sort(unique(xrf_col_res$rule_id))
+  )
+})
+
+