+++ tests/testthat/test-rule-fit-multinomial.R
@@ -175,3 +175,36 @@ test_that('non-formula method', {
   }
 
 })
+
+# ------------------------------------------------------------------------------
+
+test_that('tidy method - multi-class', {
+  skip_on_cran()
+
+  library(xrf)
+
+  xrf_cls_mod <-
+    rule_fit(trees = 3, penalty = .001) %>%
+    set_engine("xrf") %>%
+    set_mode("classification")
+
+  set.seed(1)
+  xrf_cls_fit <-
+    xrf_cls_mod %>%
+    fit(class ~ ., data = hpc_mod)
+
+  xrf_rule_res <- tidy(xrf_cls_fit)
+  expect_true(length(unique(xrf_rule_res$class)) == 4)
+  expect_true(sum(xrf_rule_res$rule == "( TRUE )") == 4)
+  expect_true(sum(xrf_rule_res$rule_id == "(Intercept)") == 4)
+
+  xrf_col_res <- tidy(xrf_cls_fit, unit = "column")
+  expect_true(length(unique(xrf_col_res$class)) == 4)
+  expect_true(sum(xrf_col_res$term == "(Intercept)") == 4)
+  expect_true(sum(xrf_col_res$rule_id == "(Intercept)") == 4)
+  expect_equal(
+    sort(unique(xrf_col_res$term)),
+    c("(Intercept)", "compounds", "input_fields", "protocol")
+  )
+})
+